refactor database connection and use database for feature toggling
